SPARK-25003: Address Reviewer Comments
